pyproject.toml
[project]
name = "ml-finance"
version = "0.1.0"
description = "Dissertation project for machine learning in finance."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-doc==0.0.3",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "beautifulsoup4==4.14.2",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "curl-cffi==0.13.0",
    "fastapi==0.121.1",
    "frozendict==2.4.7",
    "h11==0.16.0",
    "idna==3.11",
    "joblib==1.5.2",
    "lightgbm==4.6.0",
    "multitasking==0.0.12",
    "numpy==2.3.5",
    "pandas==2.3.3",
    "peewee==3.18.3",
    "platformdirs==4.5.0",
    "protobuf==6.33.1",
    "pycparser==2.23",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "requests==2.32.5",
    "scikit-learn==1.7.2",
    "scipy==1.16.3",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "starlette==0.49.3",
    "threadpoolctl==3.6.0",
    "tiingo>=0.16.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "websockets==15.0.1",
    "yfinance>=1.2.0",
]
